parser/htmlparser/tests/mochitest/html5lib_tree_construction/pending-spec-changes.dat

#data
<svg><![CDATA[]]
#errors
#document
| <html>
|   <head>
|   <body>
|     <svg svg>

#data
<svg><![CDATA[]
#errors
#document
| <html>
|   <head>
|   <body>
|     <svg svg>
